config: merge create post into tasks.py
