+++ README.md
@@ -111,7 +111,7 @@
     Run the following command to create a new post. Follow the steps to fill in all necessary information.
 
     ```bash
-    python create_post.py
+    pipenv run inv create-post
     ```
 
     Then open the newly created file under `content/posts` to finish editing the rest of the content body.
