+++ tasks.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
-import datetime
+from datetime import datetime, date
+from textwrap import dedent
 import os
 import shlex
 import shutil
@@ -12,6 +13,7 @@ from pelican import main as pelican_main
 from pelican.server import ComplexHTTPRequestHandler, RootedHTTPServer
 from pelican.settings import DEFAULT_CONFIG, get_settings_from_file
 
+# Default pelican task.py settings
 OPEN_BROWSER_ON_SERVE = True
 SETTINGS_FILE_BASE = "pelicanconf.py"
 SETTINGS = {}
@@ -26,11 +28,43 @@ CONFIG = {
     "deploy_path": SETTINGS["OUTPUT_PATH"],
     # Github Pages configuration
     "github_pages_branch": "gh-pages",
-    "commit_message": "'Publish site on {}'".format(datetime.date.today().isoformat()),
+    "commit_message": "'Publish site on {}'".format(date.today().isoformat()),
     # Host and port for `serve`
     "host": "localhost",
     "port": 8000,
 }
+
+# templates for creating a new post
+
+POST_TEMPLATE = dedent(
+    """
+    Title: {title}
+    Date: {date}
+    Modified: {modified}
+    Category: {category}
+    Tags: {tags}
+    Slug: {slug}
+    Authors: {authors}
+    Summary: {summary}
+
+    [TOC]
+
+    ---
+    <!--your content here-->
+"""
+)
+
+
+AVAILABLE_CATEGORIES = [
+    "announcement",
+    "interview",
+    "online-conference",
+    "programs",
+    "registration",
+    "sponsors",
+    "tutoiral",
+    "visiting group",
+]
 
 
 @task
@@ -191,3 +225,70 @@ def security_check(c):
         rm -rf requirements.txt
         """
     )
+
+
+def _input_date() -> str:
+    default_time = str(datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
+    return input(f"Date (default: {default_time}): ") or default_time
+
+
+def _input_category() -> str:
+    print("Select a category:")
+    for idx, cat in enumerate(AVAILABLE_CATEGORIES):
+        print(f"{idx}. {cat}")
+
+    selection = int(input("Your selection (default: 0): ") or 0)
+    if selection < 0 or selection >= len(AVAILABLE_CATEGORIES):
+        raise ValueError(f"Numbers should be within 0~{len(AVAILABLE_CATEGORIES) - 1}")
+
+    return AVAILABLE_CATEGORIES[selection]
+
+
+def _input_tags() -> str:
+    print('Type any number of tags. Enter "!" to finish\n')
+    tags = []
+    while (tag := input("-> ")) != "!":
+        tags.append(tag)
+    return ", ".join(tags)
+
+
+def _input_authors() -> str:
+    print('Specify any number of authors. Enter "!" to finish\n')
+    authors = []
+    while (author := input("-> ")) != "!":
+        authors.append(author)
+    return ", ".join(authors)
+
+
+@task
+def create_post(context) -> None:
+    """Create a new post with required metadata."""
+    print("Create a new post to pycontw-blog\n".title())
+
+    title = input("Title of the post: ")
+    date = modified = _input_date()
+    category = _input_category()
+    tags = _input_tags()
+    authors = _input_authors()
+
+    slug_title = "-".join(title.lower().split())
+    slug_date = datetime.now().strftime("%Y-%m-%d")
+    slug = f"{slug_date}-{slug_title}"
+    summary = input("Summary: ")
+
+    rendered_template = POST_TEMPLATE.format(
+        title=title,
+        date=date,
+        modified=modified,
+        category=category,
+        tags=tags,
+        authors=authors,
+        slug=slug,
+        summary=summary,
+    )
+    file_path = f"content/posts/{slug}.md"
+    with open(file_path, "w") as out:
+        out.write(rendered_template)
+
+    print(f"\nFile has already been written to {file_path}.")
+    print("Please open the file to continue editing the content. Have a nice day~")
